Rename command test files to Postfix with 'test' rather than prefix.
