+++ src/test/java/edu/hawaii/its/api/wrapper/AddMemberCommandTest.java
@@ -14,7 +14,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @ActiveProfiles("integrationTest")
 @SpringBootTest(classes = { SpringBootWebApplication.class })
-public class TestAddMemberCommand {
+public class AddMemberCommandTest {
     @Value("${groupings.api.test.grouping_many_include}")
     protected String GROUPING_INCLUDE;
