+++ src/test/java/edu/hawaii/its/api/wrapper/AddMembersCommandTest.java
@@ -14,9 +14,12 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+
 @ActiveProfiles("integrationTest")
 @SpringBootTest(classes = { SpringBootWebApplication.class })
-public class TestSubjectsCommand {
+public class AddMembersCommandTest {
+
+
     @Value("${groupings.api.test.grouping_many_include}")
     protected String GROUPING_INCLUDE;
 
@@ -28,23 +31,30 @@ public class TestSubjectsCommand {
 
     @Value("${groupings.api.success}")
     protected String SUCCESS;
-
     @Test
     public void constructorTest() {
-        assertNotNull(new SubjectsCommand(UH_NUMBERS));
-        assertNotNull(new SubjectsCommand(UH_USERNAMES));
+        AddMembersCommand addMembersCommand = new AddMembersCommand(GROUPING_INCLUDE, UH_NUMBERS);
+        assertNotNull(addMembersCommand);
+
         try {
-            new SubjectsCommand(null);
-        } catch (NullPointerException e) {
+            new AddMembersCommand(null, UH_NUMBERS);
+        }catch (NullPointerException e) {
+            assertEquals("groupPath cannot be null", e.getMessage());
+        }
+
+        try {
+            new AddMembersCommand(GROUPING_INCLUDE, null);
+        }catch (NullPointerException e) {
             assertEquals("uhIdentifiers cannot be null", e.getMessage());
         }
+
         String[] array = { "uid", "uid", null, "uid" };
         List<String> listWithNull = new ArrayList<>(Arrays.asList(array));
+
         try {
-            new SubjectsCommand(listWithNull);
-        } catch (NullPointerException e) {
+            new AddMembersCommand(GROUPING_INCLUDE, listWithNull);
+        }catch (NullPointerException e) {
             assertEquals("uhIdentifier cannot be null", e.getMessage());
         }
     }
-
 }
